Allow decimal for RXE-1-1-1

+++ CGTA_Input_Tester/ConverterLibrary/src/main/java/ca/cgta/input/model/inner/MedicationOrder.java
@@ -13,7 +13,7 @@ public class MedicationOrder {
 	public Ei myPlacerGroupNumber;
 	public Ei myPlacerOrderNumber;
 	public Visit myVisit;
-	public int myEncodedOrderQuantityNumber;
+	public double myEncodedOrderQuantityNumber;
 	public String myEncodedOrderQuantityRepeatPattern;
 	public String myEncodedOrderQuantityDuration;
 	public Date myEncodedOrderQuantityStartTime;
